# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: |
    SeedChain: direct farmer-to-customer agricultural traceability.
    Three roles only: admin, farmer, customer.
    Authentication uses an httpOnly session cookie. Every non-GET request must
    send the header `X-SeedChain-CSRF: 1`.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a farmer (pending admin approval) or a customer (active immediately)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterUserBody" }
      responses:
        "201":
          description: Registered and signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginUserBody" }
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401": { $ref: "#/components/responses/Error" }
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CurrentUser" }
        "401": { $ref: "#/components/responses/Error" }
  /me/profile:
    patch:
      operationId: updateMyProfile
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProfileBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CurrentUser" }

  # ------------------------------------------------------------------ admin users
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      parameters:
        - { name: role, in: query, schema: { $ref: "#/components/schemas/UserRole" } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/UserStatus" } }
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CurrentUser" }
  /admin/users/{id}/{action}:
    post:
      operationId: reviewUser
      tags: [admin]
      summary: Approve / reject a farmer, suspend / reactivate any non-admin user
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - name: action
          in: path
          required: true
          schema: { type: string, enum: [approve, reject, suspend, reactivate] }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReviewUserBody" }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CurrentUser" }
        "409": { $ref: "#/components/responses/Error" }
  /admin/audit-logs:
    get:
      operationId: listAuditLogs
      tags: [admin]
      parameters:
        - { name: entityType, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500 } }
      responses:
        "200":
          description: Audit log entries (newest first)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AuditLog" }
  /admin/events:
    get:
      operationId: listAllTraceEvents
      tags: [admin]
      parameters:
        - { name: eventType, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500 } }
      responses:
        "200":
          description: Traceability events across all lots (newest first)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TraceEvent" }
  /admin/scans:
    get:
      operationId: listQrScans
      tags: [admin]
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500 } }
      responses:
        "200":
          description: QR scans (newest first)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/QrScan" }

  # ------------------------------------------------------------------ farms / products
...
  /admin/integrations/{source}/run:
    post:
      operationId: runIntegration
      tags: [external]
      parameters:
        - name: source
          in: path
          required: true
          schema: { type: string, enum: [market_prices, weather] }
      responses:
        "200":
          description: Run finished
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IngestionRun" }
components:
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
        details: {}
...
    UserRole: { type: string, enum: [admin, farmer, customer] }
    UserStatus: { type: string, enum: [pending, active, rejected, suspended] }
    FarmerProfile:
      type: object
      required: [publicName]
      properties:
        publicName: { type: string }
        bio: { type: [string, "null"] }
        village: { type: [string, "null"] }
        district: { type: [string, "null"] }
        state: { type: [string, "null"] }
        verifiedAt: { type: [string, "null"], format: date-time }
        reviewNote: { type: [string, "null"] }
    CurrentUser:
      type: object
      required: [id, name, email, role, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string }
        phone: { type: [string, "null"] }
        role: { $ref: "#/components/schemas/UserRole" }
        status: { $ref: "#/components/schemas/UserStatus" }
        location: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
        farmerProfile:
          oneOf:
            - $ref: "#/components/schemas/FarmerProfile"
            - type: "null"
    FarmerProfileInput:
      type: object
      required: [publicName]
      additionalProperties: false
      properties:
        publicName: { type: string, minLength: 2, maxLength: 120 }
        bio: { type: string, maxLength: 1000 }
        village: { type: string, maxLength: 120 }
        district: { type: string, maxLength: 120 }
        state: { type: string, maxLength: 120 }
    RegisterUserBody:
      type: object
      required: [name, email, password, role]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 10, maxLength: 200 }
        role: { type: string, enum: [farmer, customer] }
        phone: { type: string, maxLength: 30 }
        location: { type: string, maxLength: 200 }
        farmerProfile: { $ref: "#/components/schemas/FarmerProfileInput" }
    LoginUserBody:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email: { type: string, maxLength: 254 }
        password: { type: string, maxLength: 200 }
    AuthResponse:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/CurrentUser" }
    UpdateProfileBody:
      type: object
      additionalProperties: false
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        phone: { type: string, maxLength: 30 }
        location: { type: string, maxLength: 200 }
        farmerProfile: { $ref: "#/components/schemas/FarmerProfileInput" }
    ReviewUserBody:
      type: object
      additionalProperties: false
      properties:
        note: { type: string, maxLength: 1000 }
    AuditLog:
      type: object
      required: [id, action, entityType, createdAt]
      properties:
        id: { type: string, format: uuid }
        userId: { type: [string, "null"] }
        userName: { type: [string, "null"] }
        role: { type: [string, "null"] }
        action: { type: string }
        entityType: { type: string }
        entityId: { type: [string, "null"] }
        before: {}
        after: {}
        requestId: { type: [string, "null"] }
        eventId: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
...
    TraceEvent:
      type: object
      required: [id, lotId, eventType, eventTime, recordedAt, source, metadata]
      properties:
        id: { type: string, format: uuid }
        lotId: { type: string, format: uuid }
        lotCode: { type: string }
        orderId: { type: [string, "null"] }
        eventType: { type: string }
        eventTime: { type: string, format: date-time }
        recordedAt: { type: string, format: date-time }
        actorUserId: { type: [string, "null"] }
        actorName: { type: [string, "null"] }
        actorRole: { type: [string, "null"] }
        location: { type: [string, "null"] }
        latitude: { type: [number, "null"] }
        longitude: { type: [number, "null"] }
        quantityBefore: { type: [number, "null"] }
        quantityChange: { type: [number, "null"] }
        quantityAfter: { type: [number, "null"] }
        status: { type: [string, "null"] }
        reason: { type: [string, "null"] }
        source: { type: string }
        clientEventId: { type: [string, "null"] }
        previousEventId: { type: [string, "null"] }
        correctsEventId: { type: [string, "null"] }
        isPublic: { type: boolean }
        metadata: { type: object, additionalProperties: true }
...
    QrScan:
      type: object
      required: [id, scannedAt, scanSource, result, signedIn]
      properties:
        id: { type: string, format: uuid }
        lotId: { type: [string, "null"] }
        lotCode: { type: [string, "null"] }
        qrVersion: { type: [integer, "null"] }
        scannedAt: { type: string, format: date-time }
        scanSource: { type: string }
        result: { type: string }
        deviceType: { type: [string, "null"] }
        signedIn: { type: boolean }
...
    IngestionRun:
      type: object
      required: [id, source, sourceEndpoint, status, recordCount, processingVersion, startedAt, requestParams]
      properties:
        id: { type: string, format: uuid }
        source: { type: string }
        sourceEndpoint: { type: string }
        requestParams: { type: object, additionalProperties: true }
        processingVersion: { type: string }
        status: { type: string, enum: [RUNNING, SUCCESS, PARTIAL, FAILED, SKIPPED] }
        recordCount: { type: integer }
        error: { type: [string, "null"] }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: [string, "null"], format: date-time }
...